Add build, orchestration, distribution features
The extension consists of the following components.

Build - Creation of OSGi bundles using the BND Maven plugin.

Instrumentation - Supported runtime environments are controlled by description bundles (feature). These are created in the following modules.

- Pure OSGi - Apache Felix, Equinox
Directory: osgi/runtime

- OSGi with infrastructure - Apache Karaf
Directory: osgi/karaf/feature

Distribution - The Apache Karaf feature is reused to create a docker image that bundles all the components of the project, including dependencies and configuration, to run the project.

- OSGi with infrastructure - Apache Karaf - Dockerized
Directory: osgi/karaf/distribution

An initial Docker Compose file is included. Project dependencies that are not part of the core application are consumed through Docker-based services. The communication and coordination of services are done over a common Docker-based network.

The access data required to create the Docker Image is transferred to the build procedure via the CI/CD System.